refactor(core): 🔨 split remaining verification methods
