+++ DomainDetective/DomainHealthCheck.Verification.cs
@@ -174,498 +174,9 @@ namespace DomainDetective {
                     processedChecks * 100d / totalChecks,
                     processedChecks,
                     totalChecks);
-            }
         }
-        /// <summary>
-        /// Analyzes a raw SMIMEA record.
-        /// </summary>
-        /// <param name="smimeaRecord">SMIMEA record text.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckSMIMEA(string smimeaRecord, CancellationToken cancellationToken = default) {
-            await SmimeaAnalysis.AnalyzeSMIMEARecords(new List<DnsAnswer> {
-                new DnsAnswer {
-                    DataRaw = smimeaRecord
-                }
-            }, _logger);
-        }
+    }
 
-
-        /// <summary>
-        /// Tests an SMTP server for open relay configuration.
-        /// </summary>
-        /// <param name="host">Target host name.</param>
-        /// <param name="port">Port to connect to. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckOpenRelayHost(string host, int port = 25, CancellationToken cancellationToken = default) {
-            ValidatePort(port);
-            await OpenRelayAnalysis.AnalyzeServer(host, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Checks if a DNS server allows recursive queries.
-        /// </summary>
-        /// <param name="host">Target server host name or IP.</param>
-        /// <param name="port">DNS port number. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckOpenResolverHost(string host, int port = 53, CancellationToken cancellationToken = default) {
-            ValidatePort(port);
-            await OpenResolverAnalysis.AnalyzeServer(host, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Checks a host for STARTTLS support.
-        /// </summary>
-        /// <param name="host">Target host name.</param>
-        /// <param name="port">Port to connect to. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckStartTlsHost(string host, int port = 25, CancellationToken cancellationToken = default) {
-            ValidatePort(port);
-            await StartTlsAnalysis.AnalyzeServer(host, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Checks a host for SMTP TLS capabilities.
-        /// </summary>
-        /// <param name="host">Target host name.</param>
-        /// <param name="port">Port to connect to. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckSmtpTlsHost(string host, int port = 25, CancellationToken cancellationToken = default) {
-            ValidatePort(port);
-            await SmtpTlsAnalysis.AnalyzeServer(host, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Checks a host for IMAP TLS capabilities.
-        /// </summary>
-        /// <param name="host">Target host name.</param>
-        /// <param name="port">Port to connect to. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckImapTlsHost(string host, int port = 143, CancellationToken cancellationToken = default) {
-            ValidatePort(port);
-            await ImapTlsAnalysis.AnalyzeServer(host, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Checks a host for POP3 TLS capabilities.
-        /// </summary>
-        /// <param name="host">Target host name.</param>
-        /// <param name="port">Port to connect to. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckPop3TlsHost(string host, int port = 110, CancellationToken cancellationToken = default) {
-            ValidatePort(port);
-            await Pop3TlsAnalysis.AnalyzeServer(host, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Retrieves the SMTP banner from a host.
-        /// </summary>
-        /// <param name="host">Target host name.</param>
-        /// <param name="port">Port to connect to. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckSmtpBannerHost(string host, int port = 25, CancellationToken cancellationToken = default) {
-            ValidatePort(port);
-            await SmtpBannerAnalysis.AnalyzeServer(host, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Measures mail server connection and banner latency.
-        /// </summary>
-        /// <param name="host">Target host name.</param>
-        /// <param name="port">Port to connect to. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckMailLatency(string host, int port = 25, CancellationToken cancellationToken = default) {
-            ValidatePort(port);
-            await MailLatencyAnalysis.AnalyzeServer(host, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Tests connectivity to common service ports on a host.
-        /// </summary>
-        /// <param name="host">Target host name.</param>
-        /// <param name="ports">Ports to check. Defaults to common services.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckPortAvailability(string host, IEnumerable<int>? ports = null, CancellationToken cancellationToken = default) {
-            var list = ports?.ToArray() ?? new[] { 25, 80, 443, 465, 587 };
-            foreach (var p in list) {
-                ValidatePort(p);
-            }
-            await PortAvailabilityAnalysis.AnalyzeServers(new[] { host }, list, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Checks a host for SNMP responses.
-        /// </summary>
-        /// <param name="host">Target host name.</param>
-        /// <param name="port">SNMP port to query.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckSnmpHost(string host, int port = 161, CancellationToken cancellationToken = default) {
-            ValidatePort(port);
-            await SnmpAnalysis.AnalyzeServer(host, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Scans a host for open TCP and UDP ports.
-        /// </summary>
-        /// <param name="host">Target host name.</param>
-        /// <param name="ports">Ports to scan. Defaults to the top 1000 ports.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task ScanPorts(string host, IEnumerable<int>? ports = null, PortScanProfile[]? profiles = null, CancellationToken cancellationToken = default, bool showProgress = true) {
-            IEnumerable<int> selected;
-            if (ports != null && ports.Any()) {
-                selected = ports;
-            } else if (profiles != null && profiles.Length > 0) {
-                selected = profiles.SelectMany(PortScanProfileDefinition.GetPorts).Distinct();
-            } else {
-                selected = PortScanProfileDefinition.DefaultPorts;
-            }
-            var list = selected.ToArray();
-            foreach (var p in list) {
-                ValidatePort(p);
-            }
-            await PortScanAnalysis.Scan(host, list, _logger, cancellationToken, showProgress);
-        }
-
-        /// <summary>Queries neighbors sharing the same IP as <paramref name="domainName"/>.</summary>
-        public async Task CheckIPNeighbors(string domainName, CancellationToken cancellationToken = default) {
-            await IPNeighborAnalysis.Analyze(domainName, _logger, cancellationToken);
-        }
-
-
-
-
-        /// <summary>
-        /// Queries random subdomains to detect wildcard DNS behavior.
-        /// </summary>
-        /// <param name="domainName">Domain to verify.</param>
-        /// <param name="sampleCount">Number of names to test.</param>
-        public async Task VerifyWildcardDns(string domainName, int sampleCount = 3) {
-            if (string.IsNullOrWhiteSpace(domainName)) {
-                throw new ArgumentNullException(nameof(domainName));
-            }
-            domainName = NormalizeDomain(domainName);
-            UpdateIsPublicSuffix(domainName);
-            WildcardDnsAnalysis.DnsConfiguration = DnsConfiguration;
-            await WildcardDnsAnalysis.Analyze(domainName, _logger, sampleCount);
-        }
-
-
-
-
-
-
-        /// <summary>
-        /// Checks all MX hosts for STARTTLS support.
-        /// </summary>
-        /// <param name="domainName">Domain whose MX records are queried.</param>
-        /// <param name="port">SMTP port to connect to. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task VerifySTARTTLS(string domainName, int port = 25, CancellationToken cancellationToken = default) {
-            if (string.IsNullOrWhiteSpace(domainName)) {
-                throw new ArgumentNullException(nameof(domainName));
-            }
-            domainName = NormalizeDomain(domainName);
-            UpdateIsPublicSuffix(domainName);
-            ValidatePort(port);
-            var mxRecordsForTls = await DnsConfiguration.QueryDNS(domainName, DnsRecordType.MX, cancellationToken: cancellationToken);
-            IEnumerable<string> tlsHosts = CertificateAnalysis.ExtractMxHosts(mxRecordsForTls);
-            await StartTlsAnalysis.AnalyzeServers(tlsHosts, new[] { port }, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Checks all MX hosts for SMTP TLS configuration.
-        /// </summary>
-        /// <param name="domainName">Domain whose MX records are queried.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task VerifySMTPTLS(string domainName, CancellationToken cancellationToken = default) {
-            if (string.IsNullOrWhiteSpace(domainName)) {
-                throw new ArgumentNullException(nameof(domainName));
-            }
-            domainName = NormalizeDomain(domainName);
-            UpdateIsPublicSuffix(domainName);
-            var mxRecordsForTls = await DnsConfiguration.QueryDNS(domainName, DnsRecordType.MX, cancellationToken: cancellationToken);
-            IEnumerable<string> tlsHosts = CertificateAnalysis.ExtractMxHosts(mxRecordsForTls);
-            await SmtpTlsAnalysis.AnalyzeServers(tlsHosts, 25, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Checks all MX hosts for IMAP TLS configuration.
-        /// </summary>
-        /// <param name="domainName">Domain whose MX records are queried.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task VerifyIMAPTLS(string domainName, CancellationToken cancellationToken = default) {
-            if (string.IsNullOrWhiteSpace(domainName)) {
-                throw new ArgumentNullException(nameof(domainName));
-            }
-            domainName = NormalizeDomain(domainName);
-            UpdateIsPublicSuffix(domainName);
-            var mxRecordsForTls = await DnsConfiguration.QueryDNS(domainName, DnsRecordType.MX, cancellationToken: cancellationToken);
-            IEnumerable<string> tlsHosts = CertificateAnalysis.ExtractMxHosts(mxRecordsForTls);
-            await ImapTlsAnalysis.AnalyzeServers(tlsHosts, 143, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Checks all MX hosts for POP3 TLS configuration.
-        /// </summary>
-        /// <param name="domainName">Domain whose MX records are queried.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task VerifyPOP3TLS(string domainName, CancellationToken cancellationToken = default) {
-            if (string.IsNullOrWhiteSpace(domainName)) {
-                throw new ArgumentNullException(nameof(domainName));
-            }
-            domainName = NormalizeDomain(domainName);
-            UpdateIsPublicSuffix(domainName);
-            var mxRecordsForTls = await DnsConfiguration.QueryDNS(domainName, DnsRecordType.MX, cancellationToken: cancellationToken);
-            IEnumerable<string> tlsHosts = CertificateAnalysis.ExtractMxHosts(mxRecordsForTls);
-            await Pop3TlsAnalysis.AnalyzeServers(tlsHosts, 110, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Collects SMTP banners from all MX hosts.
-        /// </summary>
-        /// <param name="domainName">Domain whose MX records are queried.</param>
-        /// <param name="port">SMTP port to connect to. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task VerifySMTPBanner(string domainName, int port = 25, CancellationToken cancellationToken = default) {
-            if (string.IsNullOrWhiteSpace(domainName)) {
-                throw new ArgumentNullException(nameof(domainName));
-            }
-            domainName = NormalizeDomain(domainName);
-            UpdateIsPublicSuffix(domainName);
-            ValidatePort(port);
-            var mx = await DnsConfiguration.QueryDNS(domainName, DnsRecordType.MX, cancellationToken: cancellationToken);
-            var hosts = CertificateAnalysis.ExtractMxHosts(mx);
-            await SmtpBannerAnalysis.AnalyzeServers(hosts, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Retrieves SMTP AUTH capabilities from all MX hosts.
-        /// </summary>
-        /// <param name="domainName">Domain whose MX records are queried.</param>
-        /// <param name="port">SMTP port to connect to. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task VerifySmtpAuth(string domainName, int port = 25, CancellationToken cancellationToken = default) {
-            if (string.IsNullOrWhiteSpace(domainName)) {
-                throw new ArgumentNullException(nameof(domainName));
-            }
-            domainName = NormalizeDomain(domainName);
-            UpdateIsPublicSuffix(domainName);
-            ValidatePort(port);
-            var mx = await DnsConfiguration.QueryDNS(domainName, DnsRecordType.MX, cancellationToken: cancellationToken);
-            var hosts = CertificateAnalysis.ExtractMxHosts(mx);
-            await SmtpAuthAnalysis.AnalyzeServers(hosts, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-
-
-        /// Attempts zone transfers against authoritative name servers.
-        /// </summary>
-        /// <param name="domainName">Domain to verify.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task VerifyZoneTransfer(string domainName, CancellationToken cancellationToken = default) {
-            if (string.IsNullOrWhiteSpace(domainName)) {
-                throw new ArgumentNullException(nameof(domainName));
-            }
-            domainName = NormalizeDomain(domainName);
-            UpdateIsPublicSuffix(domainName);
-            var nsRecords = await DnsConfiguration.QueryDNS(domainName, DnsRecordType.NS, cancellationToken: cancellationToken);
-            var servers = nsRecords.Select(r => r.Data.Trim('.'));
-            await ZoneTransferAnalysis.AnalyzeServers(domainName, servers, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Validates delegation information against the parent zone.
-        /// </summary>
-        /// <param name="domainName">Domain to verify.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task VerifyDelegation(string domainName, CancellationToken cancellationToken = default) {
-            if (string.IsNullOrWhiteSpace(domainName)) {
-                throw new ArgumentNullException(nameof(domainName));
-            }
-            domainName = NormalizeDomain(domainName);
-            UpdateIsPublicSuffix(domainName);
-            var ns = await DnsConfiguration.QueryDNS(domainName, DnsRecordType.NS, cancellationToken: cancellationToken);
-            await NSAnalysis.AnalyzeNsRecords(ns, _logger);
-            await NSAnalysis.AnalyzeParentDelegation(domainName, _logger);
-        }
-
-        /// <summary>
-        /// Detects dangling CNAME records for the domain.
-        /// </summary>
-        public async Task VerifyDanglingCname(string domainName, CancellationToken cancellationToken = default) {
-            domainName = NormalizeDomain(domainName);
-            DanglingCnameAnalysis = new DanglingCnameAnalysis { DnsConfiguration = DnsConfiguration };
-            await DanglingCnameAnalysis.Analyze(domainName, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Checks for CNAMEs pointing to takeover prone providers.
-        /// </summary>
-        public async Task VerifyTakeoverCname(string domainName, CancellationToken cancellationToken = default) {
-            domainName = NormalizeDomain(domainName);
-            TakeoverCnameAnalysis = new TakeoverCnameAnalysis { DnsConfiguration = DnsConfiguration };
-            await TakeoverCnameAnalysis.Analyze(domainName, _logger, cancellationToken);
-        }
-
-
-
-        private async Task VerifyReverseDnsAsync(string domainName, CancellationToken cancellationToken) {
-            var mxRecords = await DnsConfiguration.QueryDNS(domainName, DnsRecordType.MX, cancellationToken: cancellationToken);
-            var rdnsHosts = CertificateAnalysis.ExtractMxHosts(mxRecords);
-            await ReverseDnsAnalysis.AnalyzeHosts(rdnsHosts, _logger);
-        }
-
-        private async Task VerifyFcrDnsAsync(string domainName, CancellationToken cancellationToken) {
-            var mxRecords = await DnsConfiguration.QueryDNS(domainName, DnsRecordType.MX, cancellationToken: cancellationToken);
-            var rdnsHosts = CertificateAnalysis.ExtractMxHosts(mxRecords);
-            await ReverseDnsAnalysis.AnalyzeHosts(rdnsHosts, _logger);
-            await FcrDnsAnalysis.Analyze(ReverseDnsAnalysis.Results, _logger);
-        }
-
-        private async Task VerifyDaneAsync(string domainName, ServiceType[]? serviceTypes, int[]? ports, CancellationToken cancellationToken) {
-            if (ports != null && ports.Length > 0) {
-                await VerifyDANE(domainName, ports, cancellationToken);
-            } else {
-                await VerifyDANE(domainName, serviceTypes, cancellationToken);
-            }
-        }
-
-        private async Task VerifySecurityTxtAsync(string domainName, CancellationToken cancellationToken) {
-            SecurityTXTAnalysis = new SecurityTXTAnalysis();
-            await SecurityTXTAnalysis.AnalyzeSecurityTxtRecord(domainName, _logger);
-        }
-
-        private Task VerifyHpkpAsync(string domainName, CancellationToken cancellationToken) {
-            return HPKPAnalysis.AnalyzeUrl($"http://{domainName}", _logger);
-        }
-
-        private Task VerifyMessageHeaderAsync(CancellationToken cancellationToken) {
-            MessageHeaderAnalysis = CheckMessageHeaders(string.Empty, cancellationToken);
-            return Task.CompletedTask;
-        }
-
-        private Task VerifyDnsTtlAsync(string domainName, CancellationToken cancellationToken) {
-            return DnsTtlAnalysis.Analyze(domainName, _logger);
-        }
-
-        private Task VerifyFlatteningServiceAsync(string domainName, CancellationToken cancellationToken) {
-            FlatteningServiceAnalysis = new FlatteningServiceAnalysis { DnsConfiguration = DnsConfiguration };
-            return FlatteningServiceAnalysis.Analyze(domainName, _logger, cancellationToken);
-        }
-
-        private async Task<DnsAnswer[]> QueryDaneDns(string name, CancellationToken cancellationToken) {
-            if (DaneAnalysis.QueryDnsOverride != null) {
-                return await DaneAnalysis.QueryDnsOverride(name, DnsRecordType.TLSA);
-            }
-
-            return await DnsConfiguration.QueryDNS(name, DnsRecordType.TLSA, cancellationToken: cancellationToken);
-        }
-
-
-        /// <summary>
-        /// Queries SMIMEA records for an email address.
-        /// <param name="emailAddress">Email address to query.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task VerifySMIMEA(string emailAddress, CancellationToken cancellationToken = default) {
-            if (string.IsNullOrWhiteSpace(emailAddress)) {
-                throw new ArgumentNullException(nameof(emailAddress));
-            }
-
-            var name = SMIMEAAnalysis.GetQueryName(emailAddress);
-            SmimeaAnalysis = new SMIMEAAnalysis();
-            var records = await DnsConfiguration.QueryDNS(name, DnsRecordType.SMIMEA, cancellationToken: cancellationToken);
-            if (records.Any()) {
-                await SmimeaAnalysis.AnalyzeSMIMEARecords(records, _logger);
-            } else {
-                _logger.WriteWarning("No SMIMEA records found.");
-            }
-        }
-
-        /// <summary>
-        /// Verifies the certificate for a website. If no scheme is provided in <paramref name="url"/>, "https://" is assumed.
-        /// </summary>
-        /// <param name="url">Website address. If missing a scheme, "https://" will be prepended.</param>
-        /// <param name="port">Port to use for the connection. Must be between 1 and 65535.</param>
-        /// <param name="cancellationToken">Optional cancellation token.</param>
-        public async Task VerifyWebsiteCertificate(string url, int port = 443, CancellationToken cancellationToken = default) {
-            ValidatePort(port);
-            if (!url.StartsWith("http://", StringComparison.OrdinalIgnoreCase) &&
-                !url.StartsWith("https://", StringComparison.OrdinalIgnoreCase)) {
-                url = $"https://{url}";
-            }
-            await CertificateAnalysis.AnalyzeUrl(url, port, _logger, cancellationToken);
-        }
-
-        /// <summary>
-        /// Performs a basic HTTP check without enforcing HTTPS.
-        /// </summary>
-        /// <param name="domainName">Domain or host to query.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task VerifyPlainHttp(string domainName, CancellationToken cancellationToken = default) {
-            if (string.IsNullOrWhiteSpace(domainName)) {
-                throw new ArgumentNullException(nameof(domainName));
-            }
-            domainName = ValidateHostName(domainName);
-            UpdateIsPublicSuffix(domainName);
-            await HttpAnalysis.AnalyzeUrl($"http://{domainName}", false, _logger, cancellationToken: cancellationToken);
-        }
-
-        /// <summary>
-        /// Sends an ICMP echo request to a host.
-        /// </summary>
-        /// <param name="host">Target host name or address.</param>
-        /// <param name="timeout">Timeout in milliseconds.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task<PingReply> VerifyPing(string host, int timeout = 4000, CancellationToken cancellationToken = default) {
-            cancellationToken.ThrowIfCancellationRequested();
-            return await PingTraceroute.PingAsync(host, timeout, _logger);
-        }
-
-        /// <summary>
-        /// Performs a traceroute to the specified host.
-        /// </summary>
-        /// <param name="host">Target host name or address.</param>
-        /// <param name="maxHops">Maximum number of hops to probe.</param>
-        /// <param name="timeout">Timeout per hop in milliseconds.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task<IReadOnlyList<PingTraceroute.TracerouteHop>> VerifyTraceroute(string host, int maxHops = 30, int timeout = 4000, CancellationToken cancellationToken = default) {
-            cancellationToken.ThrowIfCancellationRequested();
-            return await PingTraceroute.TracerouteAsync(host, maxHops, timeout, _logger);
-        }
-
-
-
-        /// <summary>
-        /// Queries WHOIS information and IANA RDAP for a domain.
-        /// </summary>
-        /// <param name="domain">Domain name to query.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task CheckWHOIS(string domain, CancellationToken cancellationToken = default) {
-            var timeout = WhoisAnalysis.Timeout;
-            WhoisAnalysis = new WhoisAnalysis { Timeout = timeout };
-            domain = NormalizeDomain(domain);
-            UpdateIsPublicSuffix(domain);
-            await WhoisAnalysis.QueryWhoisServer(domain, cancellationToken);
-            await WhoisAnalysis.QueryIana(domain, cancellationToken);
-        }
-
-        /// <summary>
-        /// Queries RDAP information for a domain.
-        /// </summary>
-        /// <param name="domain">Domain name to query.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task QueryRDAP(string domain, CancellationToken cancellationToken = default) {
-            cancellationToken.ThrowIfCancellationRequested();
-            if (string.IsNullOrWhiteSpace(domain)) {
-                throw new ArgumentNullException(nameof(domain));
-            }
-
-            domain = NormalizeDomain(domain);
-            UpdateIsPublicSuffix(domain);
-            await RdapAnalysis.Analyze(domain, _logger, cancellationToken);
-        }
-
-        /// <summary>
         /// Creates a high level summary of key analyses.
         /// </summary>
         /// <returns>A populated <see cref="DomainSummary"/>.</returns>
